requirements.txt
# Core ML/AI libraries with versions
pika
absl-py==2.1.0
accelerate==1.0.1
datasets
huggingface-hub==0.27.1
numpy
onnx
onnxruntime-gpu
openai
pandas
scikit-learn
#scipy==1.15.1
sentence-transformers
sentencepiece
torch
torchaudio
torchvision
transformers

# Web and API tools with versions
fastapi
flask
gradio
httpx
requests
streamlit
uvicorn

# Media and image processing with versions
librosa
matplotlib
opencv-python
pillow

# Other important dependencies with versions
beautifulsoup4
lxml
pyyaml
tqdm

# Dependencies without versions (less critical or generic)
addict
aiofiles
aiohttp
aiosignal
altair
annotated-types
antlr4-python3-runtime
anyio
attrs
audioread
av
babel
backcall
basicsr
bce-python-sdk
bleach
blinker
breadability
cachetools
certifi
cffi
chardet
charset-normalizer
click
coloredlogs
contourpy
crcmod
cryptography
cycler
cython
decorator
defusedxml
deplacy
dill
distro
docopt
docx
editdistance
emoji
et_xmlfile
executing
facexlib
fastjsonschema
ffmpy
filelock
filetype
filterpy
flask-babel
flask-ngrok
flatbuffers
fonttools
frozenlist
fsspec
funasr
future
gensim
gfpgan
gitdb
gitpython
gradio_client
gradio_rangeslider
graphviz
grpcio
h11
htbuilder
httpcore
humanfriendly
hydra-core
idna
imageio
imageio-ffmpeg
ipython
itsdangerous
jaconv
jamo
jedi
jieba
jinja2
jiter
jmespath
joblib
jsonschema
jsonschema-specifications
jupyter_client
jupyter_core
jupyterlab_pygments
kaldi-native-fbank
kaldiio
kiwisolver
lazy_loader
llvmlite
lmdb
loguru
markdown
markdown-it-py
markupsafe
matplotlib-inline
mdurl
mistune
modelscope
more-itertools
moviepy
mpmath
msgpack
multidict
multiprocess
mvector
narwhals
nbclient
nbconvert
nbformat
networkx
nltk
numba
openpyxl
orjson
oss2
packaging
pandocfilters
parso
pdfminer.six
pexpect
pickleshare
pipreqs
platformdirs
pooch
proglog
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyarrow
pycountry
pycparser
pycryptodome
pydantic
pydantic_core
pydeck
pydub
pygments
pyngrok
pynndescent
pyparsing
python-dateutil
python-docx
python-dotenv
python-multipart
python-pptx
pytorch-wpe
pytz
pyzmq
rarfile
razdel
realesrgan
referencing
regex
resampy
responses
rich
rouge
rouge_score
rpds-py
ruff
safehttpx
safetensors
scikit-image
semantic-version
shellingham
six
smart-open
smmap
sniffio
soundcard
soundfile
soupsieve
soxr
st-annotated-text
stack-data
stanza
starlette
sumy
sympy
tenacity
tensorboard-data-server
tensorboardx
termcolor
threadpoolctl
tifffile
tiktoken
tinycss2
tokenizers
toml
tomlkit
torch-complex
torchinfo
tornado
traitlets
triton
typeguard
typer
typing_extensions
tzdata
ufal.udpipe
umap-learn
urllib3
visualdl
watchdog
wcwidth
webencodings
websockets
werkzeug
wrapt
xlsxwriter
xxhash
yapf
yarg
yarl
yeaudio
edge_tts